+++ openmls/src/group/public_group/diff/compute_path.rs
@@ -8,10 +8,7 @@ use crate::{
     credentials::CredentialWithKey,
     error::LibraryError,
     extensions::Extensions,
-    group::{
-        config::CryptoConfig, core_group::create_commit_params::CommitType,
-        errors::CreateCommitError,
-    },
+    group::{core_group::create_commit_params::CommitType, errors::CreateCommitError},
     key_packages::{KeyPackage, KeyPackageCreationResult},
     schedule::CommitSecret,
     treesync::{
@@ -62,10 +59,7 @@ impl<'a> PublicGroupDiff<'a> {
                 // the init key.
                 init_private_key: _,
             } = KeyPackage::builder().build_without_key_storage(
-                CryptoConfig {
-                    ciphersuite,
-                    version,
-                },
+                ciphersuite,
                 provider,
                 signer,
                 credential_with_key.ok_or(CreateCommitError::MissingCredential)?,