+++ openmls/src/treesync/tests_and_kats/tests.rs
@@ -21,9 +21,8 @@ fn that_commit_secret_is_derived_from_end_of_update_path_not_root(
     provider: &impl OpenMlsProvider,
 ) {
     let _ = provider; // get rid of warning
-    let crypto_config = CryptoConfig::with_default_version(ciphersuite);
     let mls_group_create_config = MlsGroupCreateConfig::builder()
-        .crypto_config(crypto_config)
+        .ciphersuite(ciphersuite)
         .use_ratchet_tree_extension(true)
         .build();
 
@@ -48,7 +47,7 @@ fn that_commit_secret_is_derived_from_end_of_update_path_not_root(
         );
         let key_package = KeyPackage::builder()
             .build(
-                CryptoConfig::with_default_version(ciphersuite),
+                ciphersuite,
                 &provider,
                 &credential_with_key_and_signer.signer,
                 credential_with_key_and_signer.credential_with_key.clone(),